Demo scene created and basic mechanics implemented.
Cleaned up the unity heirarchy, created a demo scene with moving
character, destroyable objects and loot collection linked to a UI
component (can take multiple players in case we decide to do
multiplayer), need to implement level changing and do weapon switching
properly.

BUGS: easy to hit objects through walls, this is due to the way we
currently have collision set on meshes, this will be sorted later on.